+++ readthedocs/doc_builder/exceptions.py
@@ -3,26 +3,11 @@
 """Exceptions raised when building documentation."""
 
 from django.utils.translation import gettext_noop
+from readthedocs.core.exceptions import ReadTheDocsBaseException
 from readthedocs.builds.constants import BUILD_STATUS_DUPLICATED
 
 
-class BuildBaseException(Exception):
-    message = None
-    status_code = None
-
-    def __init__(self, message=None, **kwargs):
-        self.status_code = kwargs.pop(
-            'status_code',
-            None,
-        ) or self.status_code or 1
-        self.message = message or self.message or self.get_default_message()
-        super().__init__(message, **kwargs)
-
-    def get_default_message(self):
-        return self.message
-
-
-class BuildAppError(BuildBaseException):
+class BuildAppError(ReadTheDocsBaseException):
     GENERIC_WITH_BUILD_ID = gettext_noop(
         'There was a problem with Read the Docs while building your documentation. '
         'Please try again later. '
@@ -31,7 +16,7 @@ class BuildAppError(BuildBaseException):
     )
 
 
-class BuildUserError(BuildBaseException):
+class BuildUserError(ReadTheDocsBaseException):
     GENERIC = gettext_noop(
         "We encountered a problem with a command while building your project. "
         "To resolve this error, double check your project configuration and installed "
